+++ ChangeLog.md
@@ -18,6 +18,7 @@
 - *Punch* is now always available unless the delver is full on gear (damage reduced to 35 to be less than the average resisted damage)
 - Adventures that end by `/give-up` no longer provide score to player profiles
 - Merchants now show how many uses equipment they're selling has
+- New Challenge: Rushing - adds a chance to not know the type of a route option
 
 #### Prophets of the Labyrinth Version 0.10.0:
 - Added **Consumables**: these resources can be used by any party member during combat at priority speed